Move serde implementation tests
