+++ src/animation/load.rs
@@ -154,225 +154,3 @@ impl Display for UnexpectedExtension {
 }
 
 impl Error for UnexpectedExtension {}
-
-#[cfg(test)]
-#[cfg(any(feature = "yaml", feature = "ron"))]
-mod tests {
-    use super::*;
-
-    use crate::{animation::Mode, Frame};
-    use std::time::Duration;
-
-    #[cfg(feature = "yaml")]
-    mod yaml {
-        use super::*;
-
-        #[test]
-        fn parse() {
-            // given
-            let content = b"
-            mode: PingPong
-            frames:
-              - index: 0 # index in the sprite sheet for that frame
-                duration: 100 # duration of the frame in milliseconds
-              - index: 1
-                duration: 100
-              - index: 2
-                duration: 120";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::PingPong);
-            assert_eq!(
-                animation.frames,
-                vec![
-                    Frame::new(0, Duration::from_millis(100)),
-                    Frame::new(1, Duration::from_millis(100)),
-                    Frame::new(2, Duration::from_millis(120)),
-                ]
-            );
-        }
-
-        #[test]
-        fn default_mode() {
-            // given
-            let content = b"
-            frames:
-              - index: 0
-                duration: 100";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::RepeatFrom(0));
-        }
-
-        #[test]
-        fn repeat() {
-            // given
-            let content = b"
-            mode: Repeat
-            frames:
-              - index: 0
-                duration: 100";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::RepeatFrom(0));
-        }
-
-        #[test]
-        fn once() {
-            // given
-            let content = b"
-            mode: Once
-            frames:
-              - index: 0
-                duration: 100";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::Once);
-        }
-
-        #[test]
-        fn repeat_from() {
-            // given
-            let content = b"
-            mode: !RepeatFrom 1
-            frames:
-              - index: 0
-                duration: 100
-              - index: 1
-                duration: 100";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::RepeatFrom(1));
-        }
-
-        #[test]
-        fn zero_duration() {
-            // given
-            let content = b"
-            frames:
-              - index: 0
-                duration: 0";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default().load("yaml", content);
-
-            // then
-            assert!(animation.is_err());
-        }
-
-        #[test]
-        fn same_duration_for_all_frames() {
-            // given
-            let content = b"
-            frame_duration: 100
-            frames:
-              - index: 0
-              - index: 1
-              - index: 2
-                duration: 200
-        ";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(
-                animation.frames,
-                vec![
-                    Frame::new(0, Duration::from_millis(100)),
-                    Frame::new(1, Duration::from_millis(100)),
-                    Frame::new(2, Duration::from_millis(200)),
-                ]
-            );
-        }
-
-        #[test]
-        fn same_duration_for_all_frames_short_hand() {
-            // given
-            let content = b"
-            frame_duration: 100
-            frames: [0, 1, 2]
-        ";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("yaml", content)
-                .unwrap();
-
-            // then
-            assert_eq!(
-                animation.frames,
-                vec![
-                    Frame::new(0, Duration::from_millis(100)),
-                    Frame::new(1, Duration::from_millis(100)),
-                    Frame::new(2, Duration::from_millis(100)),
-                ]
-            );
-        }
-    }
-
-    #[cfg(feature = "ron")]
-    mod ron {
-        use super::*;
-
-        #[test]
-        fn frames() {
-            // given
-            let content = b"
-            (
-                mode: RepeatFrom(1),
-                frames: [
-                    (
-                        index: 0, // index in the sprite sheet for that frame
-                        duration: 100, // # duration of the frame in milliseconds
-                    ),
-                    (index: 1, duration: 100),
-                    (index: 2, duration: 120),
-                ]
-            )";
-
-            // when
-            let animation = SpriteSheetAnimationLoader::default()
-                .load("ron", content)
-                .unwrap();
-
-            // then
-            assert_eq!(animation.mode, Mode::RepeatFrom(1));
-            assert_eq!(
-                animation.frames,
-                vec![
-                    Frame::new(0, Duration::from_millis(100)),
-                    Frame::new(1, Duration::from_millis(100)),
-                    Frame::new(2, Duration::from_millis(120)),
-                ]
-            );
-        }
-    }
-}
